Move pubsub core into its own module
